# docs/swagger.yaml
definitions:
  domain.Message:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact:
    email: mraisadlani@gmail.com
    name: Muhammad Rais Adlani
    url: https://gitlab.com/mraisadlani
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /DeleteFile:
    post:
      consumes:
      - application/json
      description: Delete file from FTP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: Delete File
      tags:
      - FTP Controller
  /MoveFile:
    post:
      consumes:
      - application/json
      description: Move file from FTP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: Move File
      tags:
      - FTP Controller
  /ReadFile:
    post:
      consumes:
      - application/json
      description: Read file from FTP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: Read File
      tags:
      - FTP Controller
  /RenameFile:
    post:
      consumes:
      - application/json
      description: Rename file from FTP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: Rename File
      tags:
      - FTP Controller
  /UploadFile:
    post:
      consumes:
      - application/json
      description: Upload file to FTP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Message'
      summary: Upload File
      tags:
      - FTP Controller
swagger: "2.0"